Using imports for domains and refactor structure
