+++ .meteor/packages
@@ -38,3 +38,4 @@ percolate:synced-cron
 standard-minifier-css
 standard-minifier-js
 practicalmeteor:mocha
+random
