+++ client/main.js
@@ -1,0 +1,5 @@
+import { DomainsAPI } from '../imports/api/domains.js'
+
+// Avoids breaking the project
+Domains = DomainsAPI
+
