--- lib/collections/domains.js
@@ -1,1 +1,0 @@
-Domains = new Mongo.Collection("domains")
