+++ lib/collections/reports.js
@@ -1,4 +1,9 @@
+import { Domains } from '../../imports/api/domains.js'
+
 Reports = new Mongo.Collection("reports")
+
+
+
 
 Reports.byDateRange = (domainId, startTime, endTime) => {
   return Reports.find({
