+++ server/accounts.js
@@ -1,3 +1,5 @@
+import { Domains } from '../imports/api/domains.js'
+
 Accounts.onCreateUser((options, user) => {
   const domain = options.profile.domain
   const domainId = Domains.insert({url: domain})
